codec/packet: unify `duration()` to return i64
As all other instances of timestamps (duration/pts/dts)

+++ src/codec/packet/packet.rs
@@ -131,8 +131,8 @@ impl Packet {
 	}
 
 	#[inline]
-	pub fn duration(&self) -> usize {
-		self.0.duration as usize
+	pub fn duration(&self) -> i64 {
+		self.0.duration as i64
 	}
 
 	#[inline]
